+++ CHANGELOG.md
@@ -5,11 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+
+-   Series grouping calcul & metric details page (#152)
+
 ## [0.38.0] - 2022-12-19
 
 ### Changed
 
 -   Update compute plan failed task key (#134)
+-   Use output_identifier to group series (#145)
 -   Remove deprecated parent_task_keys from TaskT (#153)
 
 ### Fixed
@@ -28,7 +35,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Open task drawer directly in cp details page (#122)
 -   Algo creation events aren't included in newsfeed anymore (#127)
 -   Renamed any tuple thing into a task thing (#129)
--   Use output_identifier to group series (#145)
 
 ### Fixed